Move analytics from unpublished package to smart-accounts-kit package
