+++ packages/smart-accounts-kit/package.json
@@ -139,7 +139,6 @@
     "@metamask/eslint-config": "14.1.0",
     "@metamask/eslint-config-nodejs": "14.0.0",
     "@metamask/eslint-config-typescript": "14.0.0",
-    "@metamask/smart-accounts-kit-analytics": "workspace:^",
     "@types/node": "^20.19.0",
     "@types/sinon": "^17.0.3",
     "@vitest/coverage-v8": "3.2.4",
@@ -152,6 +151,7 @@
     "eslint-plugin-n": "^17.10.3",
     "eslint-plugin-prettier": "^5.5.4",
     "eslint-plugin-promise": "^7.1.0",
+    "nock": "^14.0.4",
     "prettier": "^3.5.3",
     "sinon": "^18.0.0",
     "ts-node": "^10.9.2",
