+++ packages/smart-accounts-kit/src/analytics/schema.ts
