+++ packages/smart-accounts-kit/tsup.config.ts
@@ -3,8 +3,6 @@ import config from '../../shared/config/base.tsup.config';
 
 const options: Options = {
   ...config,
-  // Bundle internal analytics into the published kit (not a separate install).
-  noExternal: ['@metamask/smart-accounts-kit-analytics'],
   entry: [
     'src/index.ts',
     'src/experimental/index.ts',
